+++ server-serial.c
@@ -102,7 +102,7 @@ int is_serial_dev(const char *dev, const char *cwd, char **path_ref)
 
     if ((dev[0] != '/') && (cwd != NULL)) {
         n = snprintf(buf, sizeof(buf), "%s/%s", cwd, dev);
-        if ((n < 0) || (n >= sizeof(buf))) {
+        if ((n < 0) || ((size_t) n >= sizeof(buf))) {
             return(0);
         }
         dev = buf;
